Using Google Auth

+++ controller/api/v1/posts_api.js
@@ -2,7 +2,6 @@ const Post= require('../../../models/post');
 const Comment=require('../../../models/comment');
 
 module.exports.index = async function (req, res) {
-
     let posts = await Post.find({})
         .sort('-createdAt')
         .populate('user')
@@ -24,16 +23,29 @@ module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
-        post.remove();
-        await Comment.deleteMany({post: req.params.id});
-        return res.json(200,{
-        message: "Post andassociated comments deleted successfully!"
-        })
 
-    }catch(err){
+        if(post.user === req.user.id){
+            post.remove();
+            await Comment.deleteMany({post: req.params.id});
+
+            return res.json(200,{
+            message: "Post andassociated comments deleted successfully!"
+            })
+        }
+        else{
+            return res.json(401, {
+                message: "You can not delete this post!"
+            })
+            }
+
+    }
+    catch(err){
         console.log('*******', err);
         return res.json(500, {
             message: "Internal server error!"
         })
     }
 }
+
+
+
